refactor: extract TeamsPhoneManager.Domain layer (Stage 2)
Create the framework-free Domain class library (Clean Architecture inner
layer) and relocate provably-pure, zero-dependency code into it, keeping
namespaces stable (teams_phonemanager.*) so no using directives change:

  - ConstantsService            (Teams SKU/AppId GUIDs, naming prefixes,
                                 PowerShell module names, timeouts, etc.)
  - SwissHolidayFormulaEvaluator (Computus/Easter + cantonal date formulas)
  - ValidationResult            (extracted from ValidationService)

Domain references no projects and no packages — enforced by a new
DependencyRuleTests guard that fails the build if Avalonia, CommunityToolkit.Mvvm,
System.Management.Automation or Microsoft.Identity.Client ever leak into it.

The app project references Domain and excludes src\** from its root-level
compile glob to avoid double-compiling the layer sources. Scripts/bump-version.*
updated for the new ConstantsService.cs path.

Decision documented in docs: ObservableObject data-binding models are treated as
a Presentation concern (Dependency Rule preserved without a high-risk binding
rewrite). No PowerShell/Graph behavior touched.

Verified: solution builds Release 0/0 warnings; 105 tests pass (104 + dependency-rule).
