+++ Services/ValidationService.cs
@@ -190,22 +190,4 @@ namespace teams_phonemanager.Services
             return true;
         }
     }
-
-    public class ValidationResult
-    {
-        private readonly List<string> _errors = new();
-
-        public bool IsValid => _errors.Count == 0;
-        public IReadOnlyList<string> Errors => _errors.AsReadOnly();
-
-        public void AddError(string error)
-        {
-            _errors.Add(error);
-        }
-
-        public string GetErrorMessage()
-        {
-            return string.Join("\n", _errors);
-        }
-    }
 }
