+++ TeamsPhoneManager.slnx
@@ -1,4 +1,7 @@
 <Solution>
+  <Folder Name="/src/">
+    <Project Path="src/TeamsPhoneManager.Domain/TeamsPhoneManager.Domain.csproj" />
+  </Folder>
   <Project Path="teams-phonemanager.csproj" />
   <Project Path="teams-phonemanager.Tests/teams-phonemanager.Tests.csproj" />
 </Solution>
