+++ src/TeamsPhoneManager.Domain/ConstantsService.cs
