+++ src/TeamsPhoneManager.Domain/Holidays/SwissHolidayFormulaEvaluator.cs
